refactor to put loaders in separate file
